fix(vehicles): move form helpers to a server module, refresh model suggestions
- Extract MAX_AVATAR_BYTES, ALLOWED_AVATAR_TYPES, storeAvatarUpload, and
  parseVehicleFields from _authed.vehicles.new.tsx into a new server module
  app/models/vehicle-form.server.ts. This eliminates the top-level
  storage.server import from the client route graph, fixing the CF Workers
  build (import-protection violation).
- Also inline downscale logic in VehicleForm.tsx and
  _authed.vehicles.$vehicleId.scan.tsx instead of importing ~/lib/image.client,
  which was transitively reachable from the SSR bundle and triggered a second
  import-protection violation. Both sites use their own copy of the canvas
  resize loop with their respective maxDim/quality values.
- In VehicleForm, clear the model datalist when make or year changes so the
  next focus refetches for the new make/year combination (stale list fix).
